requirements.txt
azure.identity
httpx
openai
python-dateutil
